Incorporate `q_sort` from external source
Faced with implementation challenges, the `merge_sort` function was
added by referencing code from another student's repository
(https://github.com/yanjiew1/lab0-c). This decision was made to
address difficulties in the implementation process and to ensure the
efficacy of the sorting functionality. By acknowledging the use of
external contributions, this update underscores the collaborative
nature of open source development and the valuable role of community
support in overcoming technical challenges.