Optimize get_midpoint function
Previously, the get_midpoint function was designed using a fast and slow pointerstrategy for singly linked lists.It required the fast pointer to traverse the
entire list (n steps) and the slow pointer to traverse half of i (n/2 steps),
, resulting in a total of 3n/2 steps to find the middle node.

This update modifies the approach for a doubly linked circular list, where
traversing n+1 times suffices to find the midpoint. The improvement is achieved by moving two pointers from opposite ends towards the center. This method
reduces the number of steps required to identify the midpoint in such structure.